filter and some logique

+++ app/Http/Controllers/ActivityController.php
@@ -10,6 +10,7 @@ use App\Models\Centre;
 use App\Models\Participer;
 use App\Models\Type;
 use App\Repositories\Interfaces\ActivityRepositoryInterface;
+use Carbon\Carbon;
 use Exception;
 use Illuminate\Http\Request;
 
@@ -21,10 +22,29 @@ class ActivityController extends Controller
     {
         $this->activityRepository = $activityRepository;
     }
-    public function index()
+    public function index(Request $request)
     {
-        $activities = Activity::all();
-        return view('home', ['activities' => $activities]);
+        $activities = [];
+
+        if ($request->has('categorie_name') && $request->categorie_name) {
+            $categorie = Categorie::where('name', $request['categorie_name'])->first();
+            $activities = Activity::where('categorie_id', $categorie->id)->where('date_debut', '>=',  Carbon::now())->get();
+        } else {
+            $activities = Activity::where('date_debut', '>=',  Carbon::now())->get();
+        }
+        
+        $categories = Categorie::all();
+
+        return view('home', [
+            'activities' => $activities,
+            'categories' => $categories
+        ]);
+    }
+
+    public function GetSameCityActivities()
+    {
+        $user = auth()->user();
+        return $user;
     }
 
     public function create()
@@ -148,11 +168,20 @@ class ActivityController extends Controller
 
     public function delete($id)
     {
+
+
+
+
         $user = auth()->user()->id;
         $activity = Activity::find($id);
+        $date_debut = Carbon::parse($activity->date_debut);
+        $now = Carbon::now();
+        if ($now->diffInHours($date_debut) < 24) {
+            return redirect()->route('userActivities')->with('error', 'the activity is about to begin , can not leave it');
+        }
         $activity->users()->detach($user);
         $activity->participants--;
         $activity->save();
-        return redirect()->back();
+        return redirect()->route('userActivities')->with('success', 'you left that activity');
     }
 }
